requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
constantly==23.10.4
crispy-bootstrap5==2023.10
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.5
django-allauth==65.3.1
django-ckeditor==6.7.0
django-cleanup==8.1.0
django-cors-headers==4.6.0
django-countries==7.5.1
django-crispy-forms==2.1
django-filter==24.3
django-js-asset==2.2.0
django-taggit==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.5.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==0.6.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-core==0.3.31
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.3.1
Levenshtein==0.26.1
looseversion==1.3.0
lxml==5.3.1
Markdown==3.7
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
oauthlib==3.2.2
ollama==0.4.7
openai==0.28.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
proto-plus==1.25.0
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
pytesseract==0.3.13
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.11.0
rdflib==6.3.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.15.2
service-identity==24.2.0
setuptools==75.6.0
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.46.0
stripe==11.4.1
tenacity==9.0.0
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
